added the function into the new utils folder and also changed where the function is getting called form in the job
