+++ utils/estimate_filled_posts_by_job_role_utils/utils.py
@@ -71,3 +71,30 @@ def count_job_role_per_establishment_as_columns(
         df = df.withColumnRenamed(column, f"job_role_count_{column}")
 
     return df
+
+def merge_dataframes(posts_df: DataFrame, workers_df: DataFrame) -> DataFrame:
+    """
+    Joining the IndCQC Estimates dataframe (Left Table) and the ASCWDS Worker DataFrame (Right Table) together.
+    All the columns from IndCQC and the Job Count columns from ASCWDS.
+
+    Args:
+        posts_df (DataFrame): A dataframe containing cleaned IndCQC workplace Data.
+        workers_df (DataFrame): ASC-WDS worker dataframe grouped to include columns with job role counts per job role. .
+
+    Returns:
+        DataFrame: The IndCQC DataFrame merged to include job role count columns.
+    """
+
+    joined_df = posts_df.join(
+        workers_df,
+        (posts_df[IndCQC.establishment_id] == workers_df[AWKClean.establishment_id])
+        & (
+            posts_df[IndCQC.ascwds_workplace_import_date]
+            == workers_df[AWKClean.ascwds_worker_import_date]
+        ),
+        "left",
+    )
+
+    result_df = joined_df.select(posts_df, workers_df[list_of_job_roles])
+
+    return result_df
